pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ag-enforcement-tracker"
version = "0.1.0"
description = "Data pipeline for scraping, extracting, and normalizing state AG enforcement actions"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1.0",
    "selectolax>=0.3,<1.0",
    "lxml>=5.0,<6.0",
    "pydantic>=2.0,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "click>=8.0,<9.0",
    "rich>=13.0,<14.0",
    "thefuzz[speedup]>=0.22,<1.0",
    "dateparser>=1.2,<2.0",
    "pyyaml>=6.0,<7.0",
    "aiosqlite>=0.20,<1.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.29,<1.0",
]
dashboard = [
    "streamlit>=1.32,<2.0",
    "plotly>=5.0,<6.0",
    "pandas>=2.0,<3.0",
]
llm = [
    "anthropic>=0.25,<1.0",
]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.3,<1.0",
]
all = [
    "ag-enforcement-tracker[api,dashboard,llm,dev]",
]

[project.scripts]
ag-tracker = "src.cli:cli"

[tool.setuptools.packages.find]
include = ["src*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "network: marks tests requiring network access",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
